+++ NEWS.md
@@ -1,3 +1,12 @@
+# searcher 0.0.5.1
+
+## Features
+
+- Added search portal:
+    - `search_ecosia()`: Searches on [Ecosia](https://www.ecosia.org/). 
+      ([#31](https://github.com/r-assist/searcher/issues/31), 
+       [#32](https://github.com/r-assist/searcher/pull/32))
+
 # searcher 0.0.5
 
 ## Features